+++ enderchest/instance.py
@@ -1,10 +1,7 @@
 """Specification of a Minecraft instance"""
-import json
-from configparser import ConfigParser, ParsingError, SectionProxy
+from configparser import SectionProxy
 from pathlib import Path
 from typing import NamedTuple
-
-from .loggers import GATHER_LOGGER
 
 
 class InstanceSpec(NamedTuple):
@@ -57,202 +54,3 @@ class InstanceSpec(NamedTuple):
             section.get("modloader", None),
             tuple(section.get("tags", "").strip().split()),
         )
-
-
-def gather_metadata_for_official_instance(
-    minecraft_folder: Path, name: str = "official"
-) -> InstanceSpec:
-    """Parse files to generate metadata for an official Minecraft installation
-
-    Parameters
-    ----------
-    minecraft_folder : Path
-        The path to the installation's .minecraft folder
-    name : str, optional
-        A name or alias to give to the instance. If None is provided, the
-        default name is "official"
-
-    Returns
-    -------
-    InstanceSpec
-        The metadata for this instance
-
-    Raises
-    ------
-    ValueError
-        If this is not a valid official Minecraft installation
-
-    Notes
-    -----
-    This method will always consider this instance to be vanilla, with no
-    modloader. If a Forge or Fabric executable is installed inside this
-    instance, the precise name of that version of that modded minecraft
-    will be included in the version list.
-    """
-    launcher_profile_file = minecraft_folder / "launcher_profiles.json"
-    try:
-        with launcher_profile_file.open() as lp_json:
-            launcher_profiles = json.load(lp_json)
-        raw_versions: list[str] = [
-            profile["lastVersionId"]
-            for profile in launcher_profiles["profiles"].values()
-        ]
-    except FileNotFoundError as no_json:
-        raise ValueError(f"Could not find {launcher_profile_file}") from no_json
-    except json.JSONDecodeError as bad_json:
-        raise ValueError(
-            f"{launcher_profile_file} is corrupt and could not be parsed"
-        ) from bad_json
-    except KeyError as weird_json:
-        raise ValueError(
-            f"Could not parse metadata from {launcher_profile_file}"
-        ) from weird_json
-
-    version_manifest_file = minecraft_folder / "versions" / "version_manifest_v2.json"
-    try:
-        with version_manifest_file.open() as vm_json:
-            version_lookup: dict[str, str] = json.load(vm_json)["latest"]
-    except FileNotFoundError as no_json:
-        raise ValueError(f"Could not find {version_manifest_file}") from no_json
-    except json.JSONDecodeError as bad_json:
-        raise ValueError(
-            f"{version_manifest_file} is corrupt and could not be parsed"
-        ) from bad_json
-    except KeyError as weird_json:
-        GATHER_LOGGER.warn(
-            f"{version_manifest_file} has no latest-version lookup."
-            "\nPlease check the parsed metadata to ensure that it's accurate.",
-        )
-        version_lookup = {}
-
-    versions: list[str] = []
-    tags: list[str] = ["vanilla"]
-    for version in raw_versions:
-        if version.startswith("latest-"):
-            mapped_version = version_lookup.get(version[len("latest-") :])
-            if mapped_version is not None:
-                versions.append(mapped_version)
-                tags.append(version)
-                continue
-        versions.append(version)
-
-    return InstanceSpec(name, minecraft_folder, tuple(versions), None, tuple(tags))
-
-
-def gather_metadata_for_mmc_instance(
-    minecraft_folder: Path, instgroups_file: Path | None = None
-) -> InstanceSpec:
-    """Parse files to generate metadata for a MultiMC-like instance
-
-    Parameters
-    ----------
-    minecraft_folder : Path
-        The path to the installation's .minecraft folder
-    instgroups_file : Path
-        The path to instgroups.json. If None is provided, this method will
-        look for it two directories up from the minecraft folder
-
-    Returns
-    -------
-    InstanceSpec
-        The metadata for this instance
-
-    Raises
-    ------
-    ValueError
-        If this is not a valid MMC-like Minecraft instance
-
-    Notes
-    -----
-    If this method is failing to find the appropriate files, you may want
-    to try ensuring that minecraft_folder is an absolute path.
-    """
-    mmc_pack_file = minecraft_folder.parent / "mmc-pack.json"
-    try:
-        with mmc_pack_file.open() as mmc_json:
-            components: list[dict] = json.load(mmc_json)["components"]
-
-        version: str | None = None
-        modloader: str | None = None
-
-        for component in components:
-            match component.get("uid"), component.get("cachedName", ""):
-                case "net.minecraft", _:
-                    version = component["cachedVersion"]
-                case "net.fabricmc.fabric-loader", _:
-                    modloader = "Fabric Loader"
-                case "org.quiltmc.quilt-loader", _:
-                    modloader = "Quilt Loader"
-                case ("net.minecraftforge", _) | (_, "Forge"):
-                    modloader = "Forge"
-                case _, name if name.endswith("oader"):
-                    modloader = name
-                case _:
-                    continue
-        if version is None:
-            raise KeyError("Could not find a net.minecraft component")
-    except FileNotFoundError as no_json:
-        raise ValueError(f"Could not find {mmc_pack_file}") from no_json
-    except json.JSONDecodeError as bad_json:
-        raise ValueError(
-            f"{mmc_pack_file} is corrupt and could not be parsed"
-        ) from bad_json
-    except KeyError as weird_json:
-        raise ValueError(
-            f"Could not parse metadata from {mmc_pack_file}"
-        ) from weird_json
-
-    name = minecraft_folder.parent.name
-
-    tags: list[str] = []
-
-    if name == "":
-        GATHER_LOGGER.warn(
-            "Could not resolve the name of the parent folder"
-            " and thus could not load tags."
-        )
-    else:
-        instgroups_file = (
-            instgroups_file or minecraft_folder.parent.parent / "instgroups.json"
-        )
-
-        try:
-            with instgroups_file.open() as groups_json:
-                groups: dict[str, dict] = json.load(groups_json)["groups"]
-            for tag, metadata in groups.items():
-                # interestingly this comes from the folder name, not the actual name
-                if name in metadata.get("instances", ()):
-                    tags.append(tag)
-
-        except FileNotFoundError as no_json:
-            GATHER_LOGGER.warn(
-                f"Could not find {instgroups_file} and thus could not load tags"
-            )
-        except json.JSONDecodeError as bad_json:
-            GATHER_LOGGER.warn(
-                f"{instgroups_file} is corrupt and could not be parsed for tags"
-            )
-        except KeyError as weird_json:
-            GATHER_LOGGER.warn(f"Could not parse tags from {instgroups_file}")
-
-    instance_cfg = minecraft_folder.parent / "instance.cfg"
-
-    try:
-        parser = ConfigParser(allow_no_value=True)
-        parser.read_string("[instance]\n" + instance_cfg.read_text())
-        name = parser["instance"]["name"]
-    except FileNotFoundError as no_cfg:
-        GATHER_LOGGER.warn(
-            f"Could not find {instance_cfg} and thus could not load the instance name"
-        )
-    except ParsingError as no_cfg:
-        GATHER_LOGGER.warn(
-            f"{instance_cfg} is corrupt and could not be parsed the instance name"
-        )
-    except KeyError as weird_json:
-        GATHER_LOGGER.warn(f"Could not parse instance name from {instance_cfg}")
-
-    if name == "":
-        raise ValueError("Could not determine the name of the instance.")
-
-    return InstanceSpec(name, minecraft_folder, (version,), modloader, tuple(tags))
